Update Quadratic.java
A whole bunch of changes mostly in solveQuadratic().  Mostly having checks for situations that involve the discriminant. Lets say the a = 0 and b = 2 and c = 9.  The discriminant would be 2, which would satisfy hasDistinctRealRoots(). However if you did the quadratic equation you would get 0, which would be wrong. So I have a check for if a == 0. which based on different situations of a being positive/negative or c being positive/negative that will give an appropriate answer. I tested it and so far everything seems fine. I think that we need to just tidy up how the roots are presented but like I tested a = 2, b = 0, c = 9 and it gives a correct answer but it isn't formatted for a certain number of decimals. The equation prints fine but like I said the answers for decimals are not formatted yet.  So formatting and just testing all the methods is whats left but I feel most of the hard work is done.